Simplify enemy movement: only vertical and horizontal
Add test cases

+++ src/uni_defense/logic/world/MovableObject.java
@@ -2,40 +2,40 @@ package uni_defense.logic.world;
 
 public abstract class MovableObject {
 
-    private float x;
+    private double x;
     
-    private float y;
+    private double y;
 
-    public MovableObject(float x, float y) {
+    public MovableObject(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
     
-    public void move(float dx, float dy) {
+    public void move(double dx, double dy) {
         this.x += dx;
         this.y += dy;
     }
     
     /**
-     * @param dtime The time since the last step in nanseconds.
+     * @param dtime The time since the last step in milliseconds.
      */
-    public void step(long dtime) {
+    public void step(double dtime) {
     }
     
 }
